Add BUCK file parser and buck rule patcher

+++ src/commands/migrate.rs
@@ -4,6 +4,7 @@ use cargo_metadata::{MetadataCommand, Node, camino::Utf8PathBuf};
 use clap::Parser;
 
 use crate::{
+    buck::{parse_buck_file, patch_buck_rules},
     buckify::{
         buckify_dep_node, buckify_root_node, check_dep_target, gen_buck_content, vendor_package,
     },
@@ -12,9 +13,6 @@ use crate::{
 
 #[derive(Parser, Debug)]
 pub struct MigrateArgs {
-    /// generate minimal BUCK rules
-    #[arg(long, default_value = "false")]
-    pub minimal: bool,
     /// override existing source files
     #[arg(long, default_value = "false")]
     #[clap(name = "override")]
@@ -53,7 +51,7 @@ pub fn execute(args: &MigrateArgs) {
     let buck_path = Utf8PathBuf::from(cwd.to_str().unwrap()).join("BUCK");
 
     // Generate BUCK rules
-    let buck_rules = buckify_root_node(root_node, &packages_map, args.minimal);
+    let buck_rules = buckify_root_node(root_node, &packages_map);
 
     // Generate the BUCK file
     let buck_content = gen_buck_content(&buck_rules);
@@ -61,7 +59,7 @@ pub fn execute(args: &MigrateArgs) {
 
     // Process dep nodes
     let mut queue: VecDeque<Node> = VecDeque::new();
-    let mut visited: HashSet<&str> = HashSet::new();
+    let mut visited: HashSet<String> = HashSet::new();
 
     for dep in &root_node.deps {
         if let Some(node) = nodes_map.get(&dep.pkg) {
@@ -71,17 +69,35 @@ pub fn execute(args: &MigrateArgs) {
 
     while let Some(node) = queue.pop_front() {
         let package = packages_map.get(&node.id).unwrap().to_owned();
+
+        // Skip already visited nodes
+        if visited.contains(node.id.repr.as_str()) {
+            continue;
+        }
+        visited.insert(node.id.repr.to_owned());
+
+        // Skip first-party dependencies
+        if package.source.is_none() {
+            continue;
+        }
+
         println!("Processing package: {}", package.name);
 
         // Vendor package sources
         let vendor_path = vendor_package(&package, args._override);
 
-        // Create BUCK file
-        let buck_path = vendor_path.join("BUCK");
-        std::fs::File::create(&buck_path).expect("Failed to create BUCK file");
-
         // Generate BUCK rules
-        let buck_rules = buckify_dep_node(&node, &packages_map, args.minimal);
+        let mut buck_rules = buckify_dep_node(&node, &packages_map);
+
+        // Patch BUCK Rules
+        let buck_path = vendor_path.join("BUCK");
+        if buck_path.exists() && !args._override {
+            let existing_rules =
+                parse_buck_file(&buck_path).expect("Failed to parse existing BUCK file");
+            patch_buck_rules(&existing_rules, &mut buck_rules);
+        } else {
+            std::fs::File::create(&buck_path).expect("Failed to create BUCK file");
+        }
 
         // Generate the BUCK file
         let buck_content = gen_buck_content(&buck_rules);
@@ -92,7 +108,7 @@ pub fn execute(args: &MigrateArgs) {
                 continue; // Skip dependencies that do not match the target
             }
             if let Some(next_node) = nodes_map.get(&dep.pkg)
-                && visited.insert(&next_node.id.repr)
+                && !visited.contains(next_node.id.repr.as_str())
             {
                 queue.push_back(next_node.clone());
             }
